requirements.txt
python-telegram-bot==13.17
firebase-admin==6.0.0
pytz
